refactor: auto-split test_streaming.py
